refactor: consolidate related small modules to reduce file scatter
- Merge recording-badge.ts into recording-utils.ts (badge timer + slug
  generation are both recording helpers used by the same callers)
- Merge message-forwarding.ts into window-message-listener.ts (forwarding
  and listening share the same data flow)
- Update imports in recording-listeners.ts and recording.ts
- Old modules now re-export for backward compatibility
